+++ tests/testdata/bidsmap_example_out.yaml
@@ -19,7 +19,7 @@ Options:
 # --------------------------------------------------------------------------------
 # General options
 # --------------------------------------------------------------------------------
-  version: 1.6                                    # BIDScoin version (should correspond with the version in ../bidscoin/version.txt)
+  version: 2.0                                    # BIDScoin version (should correspond with the version in ../bidscoin/version.txt)
   dcm2niix:                                       # See dcm2niix -h and https://www.nitrc.org/plugins/mwiki/index.php/dcm2nii:MainPage#General_Usage for more info
     path: module add dcm2niix;                    # Command to set the path to dcm2niix (note the semi-colon), e.g. module add dcm2niix/1.0.20180622; or PATH=/opt/dcm2niix/bin:$PATH; or /opt/dcm2niix/bin/ or '"C:\Program Files\dcm2niix\"' (note the quotes to deal with the whitespace)
     args: -b y -z y -i n                          # Argument string that is passed to dcm2niix. Tip: SPM users may want to use '-z n' (which produces unzipped nifti's, see dcm2niix -h for more information)
@@ -32,38 +32,13 @@ DICOM:
   participant_label:                              # A <<DICOM attribute>> that is used as participant_label instead of the subject-label from the sourcefolder
   session_label:                                  # A <<DICOM attribute>> that is used as session_label instead of the session-label from the sourcefolder
   anat:       # ----------------------- All anatomical series --------------------
-  - attributes:                                   # From: M:\bidscoin\raw\sub-P002\ses-mri01\04_t1_mpr_sag_p2_iso_1\M109.MR.WUR_BRAIN_ADHD.0004.0001.2018.03.01.13.05.10.140625.104364139.IMA
-      SeriesDescription: t1_mpr_sag_p2_iso_1
-      SequenceVariant: "['SP', 'MP']"
-      SequenceName: '*tfl3d1_ns'
-      ScanningSequence: "['GR', 'IR']"
-      MRAcquisitionType: 3D
-      FlipAngle: '9'
-      EchoNumbers: 1
-      EchoTime: '2.26'
-      RepetitionTime: '1900'
-      ImageType: "['ORIGINAL', 'PRIMARY', 'M', 'ND', 'NORM']"
-      ProtocolName: t1_mpr_sag_p2_iso_1
-      PhaseEncodingDirection:
-    bids:
-      acq_label: t1mprsagp2iso1
-      rec_label:
-      ce_label:
-      task_label:
-      echo_index:
-      dir_label:
-      run_index: <<1>>
-      suffix:
-      mod_label:
-      modality_label:
-    provenance: M:\bidscoin\raw\sub-P002\ses-mri01\04_t1_mpr_sag_p2_iso_1\M109.MR.WUR_BRAIN_ADHD.0004.0001.2018.03.01.13.05.10.140625.104364139.IMA
   func:       # ----------------------- All functional series --------------------
   dwi:        # ----------------------- All diffusion series ---------------------
   fmap:       # ----------------------- All fieldmap series ----------------------
   beh:        # ----------------------- All behavioural data ---------------------
   pet:        # ----------------------- All PET series ---------------------------
   extra_data: # ----------------------- All extra data ---------------------------
-  - attributes:                                   # From: M:\bidscoin\raw\sub-P002\ses-mri01\02_localizer AANGEPAST 11 SLICES\M109.MR.WUR_BRAIN_ADHD.0002.0001.2018.03.01.13.05.10.140625.104357083.IMA
+  - attributes:
       SeriesDescription: localizer AANGEPAST 11 SLICES
       SequenceVariant: "['SP', 'OSP']"
       SequenceName: '*fl2d1'
@@ -87,8 +62,9 @@ DICOM:
       suffix:
       mod_label:
       modality_label:
-    provenance: M:\bidscoin\raw\sub-P002\ses-mri01\02_localizer AANGEPAST 11 SLICES\M109.MR.WUR_BRAIN_ADHD.0002.0001.2018.03.01.13.05.10.140625.104357083.IMA
-  - attributes:                                   # From: M:\bidscoin\raw\sub-P002\ses-mri01\03_Stoptaak_ep2d_bold_nomoco\M109.MR.WUR_BRAIN_ADHD.0003.0001.2018.03.01.13.05.10.140625.104359017.IMA
+    provenance: /home/tg/rutvdee/bidscoin/raw/sub-P002/ses-mri01/02_localizer AANGEPAST
+      11 SLICES/M109.MR.WUR_BRAIN_ADHD.0002.0001.2018.03.01.13.05.10.140625.104357083.IMA
+  - attributes:
       SeriesDescription: Stoptaak_ep2d_bold_nomoco
       SequenceVariant: SK
       SequenceName: '*epfid2d1_64'
@@ -112,8 +88,33 @@ DICOM:
       suffix:
       mod_label:
       modality_label:
-    provenance: M:\bidscoin\raw\sub-P002\ses-mri01\03_Stoptaak_ep2d_bold_nomoco\M109.MR.WUR_BRAIN_ADHD.0003.0001.2018.03.01.13.05.10.140625.104359017.IMA
-  - attributes:                                   # From: M:\bidscoin\raw\sub-P002\ses-mri01\05_Flanker_ep2d_bold_nomoco\M109.MR.WUR_BRAIN_ADHD.0005.0001.2018.03.01.13.05.10.140625.104368237.IMA
+    provenance: /home/tg/rutvdee/bidscoin/raw/sub-P002/ses-mri01/03_Stoptaak_ep2d_bold_nomoco/M109.MR.WUR_BRAIN_ADHD.0003.0001.2018.03.01.13.05.10.140625.104359017.IMA
+  - attributes:
+      SeriesDescription: t1_mpr_sag_p2_iso_1
+      SequenceVariant: "['SP', 'MP']"
+      SequenceName: '*tfl3d1_ns'
+      ScanningSequence: "['GR', 'IR']"
+      MRAcquisitionType: 3D
+      FlipAngle: '9'
+      EchoNumbers: 1
+      EchoTime: '2.26'
+      RepetitionTime: '1900'
+      ImageType: "['ORIGINAL', 'PRIMARY', 'M', 'ND', 'NORM']"
+      ProtocolName: t1_mpr_sag_p2_iso_1
+      PhaseEncodingDirection:
+    bids:
+      acq_label: t1mprsagp2iso1
+      rec_label:
+      ce_label:
+      task_label:
+      echo_index:
+      dir_label:
+      run_index: <<1>>
+      suffix:
+      mod_label:
+      modality_label:
+    provenance: /home/tg/rutvdee/bidscoin/raw/sub-P002/ses-mri01/04_t1_mpr_sag_p2_iso_1/M109.MR.WUR_BRAIN_ADHD.0004.0001.2018.03.01.13.05.10.140625.104364139.IMA
+  - attributes:
       SeriesDescription: Flanker_ep2d_bold_nomoco
       SequenceVariant: SK
       SequenceName: '*epfid2d1_64'
@@ -137,8 +138,8 @@ DICOM:
       suffix:
       mod_label:
       modality_label:
-    provenance: M:\bidscoin\raw\sub-P002\ses-mri01\05_Flanker_ep2d_bold_nomoco\M109.MR.WUR_BRAIN_ADHD.0005.0001.2018.03.01.13.05.10.140625.104368237.IMA
-  - attributes:                                   # From: M:\bidscoin\raw\sub-P002\ses-mri02\07_t1_fl3d_sag_p3_iso_1\M005.MR.WUR_BRAIN_ADHD.0007.0001.2018.04.12.13.00.48.734375.108749947.IMA
+    provenance: /home/tg/rutvdee/bidscoin/raw/sub-P002/ses-mri01/05_Flanker_ep2d_bold_nomoco/M109.MR.WUR_BRAIN_ADHD.0005.0001.2018.03.01.13.05.10.140625.104368237.IMA
+  - attributes:
       SeriesDescription: t1_fl3d_sag_p3_iso_1
       SequenceVariant: SP
       SequenceName: '*fl3d1_ns'
@@ -162,7 +163,7 @@ DICOM:
       suffix:
       mod_label:
       modality_label:
-    provenance: M:\bidscoin\raw\sub-P002\ses-mri02\07_t1_fl3d_sag_p3_iso_1\M005.MR.WUR_BRAIN_ADHD.0007.0001.2018.04.12.13.00.48.734375.108749947.IMA
+    provenance: /home/tg/rutvdee/bidscoin/raw/sub-P002/ses-mri02/07_t1_fl3d_sag_p3_iso_1/M005.MR.WUR_BRAIN_ADHD.0007.0001.2018.04.12.13.00.48.734375.108749947.IMA
 PAR:
 # --------------------------------------------------------------------------------
 # PAR key-value heuristics (Philips PAR fields that are mapped to the BIDS labels)
