+++ tests/testdata/bidsmap_sample_example.yaml
@@ -1,0 +1,25 @@
+attributes:
+  SeriesDescription: localizer AANGEPAST 11 SLICES
+  SequenceVariant: "['SP', 'OSP']"
+  SequenceName: '*fl2d1'
+  ScanningSequence: GR
+  MRAcquisitionType: 2D
+  FlipAngle: '20'
+  EchoNumbers: 1
+  EchoTime: '4'
+  RepetitionTime: '8.6'
+  ImageType: "['ORIGINAL', 'PRIMARY', 'M', 'NORM', 'DIS2D']"
+  ProtocolName: localizer AANGEPAST 11 SLICES
+  PhaseEncodingDirection:
+bids:
+  acq_label: localizerAANGEPAST11SLICES
+  rec_label:
+  ce_label:
+  task_label:
+  echo_index:
+  dir_label:
+  run_index: <<1>>
+  suffix:
+  mod_label:
+  modality_label:
+provenance: /home/tg/rutvdee/bidscoin/raw/sub-P002/ses-mri01/02_localizer AANGEPAST 11 SLICES/M109.MR.WUR_BRAIN_ADHD.0002.0001.2018.03.01.13.05.10.140625.104357083.IMA
